explicit casts to avoid warnings + pragma warning (win32)

+++ compiler/parser/sourcefetcher.cpp
@@ -681,7 +681,7 @@ const char* http_strerror()
 			convertedError[0] = 0;	/* Start off with NULL */
 			stringIndex = strstr(originalError, "%d");
 			strncat(convertedError, originalError,	/* Copy up to %d */
-            labs(stringIndex - originalError));
+            labs(long(stringIndex - originalError)));
 			sprintf(&convertedError[strlen(convertedError)], "%d", errorInt);
 			stringIndex += 2;	/* Skip past the %d */
 			strcat(convertedError, stringIndex);
